+++ polymer/eduke32/source/osdfuncs.h
@@ -29,7 +29,7 @@ void GAME_onshowosd(int32_t shown);
 void GAME_clearbackground(int32_t numcols, int32_t numrows);
 
 extern int32_t osdhightile;
-extern double osdscale;
+extern float osdscale;
 
 #define OSDCHAR_WIDTH 8
